+++ packages/core-js/modules/esnext.uint8-array.to-hex.js
@@ -1,0 +1,22 @@
+'use strict';
+var $ = require('../internals/export');
+var global = require('../internals/global');
+var uncurryThis = require('../internals/function-uncurry-this');
+var anUint8Array = require('../internals/an-uint8-array');
+
+var Uint8Array = global.Uint8Array;
+var numberToString = uncurryThis(1.0.toString);
+
+// `Uint8Array.prototype.toHex` method
+// https://github.com/tc39/proposal-arraybuffer-base64
+if (Uint8Array) $({ target: 'Uint8Array', proto: true, forced: true }, {
+  toHex: function toHex() {
+    anUint8Array(this);
+    var result = '';
+    for (var i = 0, length = this.length; i < length; i++) {
+      var hex = numberToString(this[i], 16);
+      result += hex.length === 1 ? '0' + hex : hex;
+    }
+    return result;
+  }
+});