+++ tests/compat/tests.js
@@ -1830,6 +1830,18 @@ GLOBAL.tests = {
   'esnext.typed-array.unique-by': function () {
     return Int8Array.prototype.uniqueBy;
   },
+  'esnext.uint8-array.from-base64': function () {
+    return Uint8Array.fromBase64;
+  },
+  'esnext.uint8-array.from-hex': function () {
+    return Uint8Array.fromHex;
+  },
+  'esnext.uint8-array.to-base64': function () {
+    return Uint8Array.prototype.toBase64;
+  },
+  'esnext.uint8-array.to-hex': function () {
+    return Uint8Array.prototype.toHex;
+  },
   'esnext.weak-map.delete-all': function () {
     return WeakMap.prototype.deleteAll;
   },